apps/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
aiosqlite
pydantic>=2.0
pydantic-settings>=2.0
email-validator
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
cryptography
asyncssh
ansible-runner
redis
httpx
pytest
pytest-asyncio
